DBZ-8035 Move activity classes to distinct package
