+++ debezium-core/src/test/java/io/debezium/metrics/activity/ActivityMonitoringMeterTest.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.debezium.transforms;
+package io.debezium.metrics.activity;
 
 import static org.assertj.core.api.Assertions.assertThat;
